+++ src/main/java/com/simplon/dvdstore/controllers/dvd/DvdStoreDtoWithOption.java
@@ -1,6 +1,4 @@
 package com.simplon.dvdstore.controllers.dvd;
 
-import java.math.BigDecimal;
-import java.util.Optional;
 
-public record DvdStoreDtoWithOption(Long id, String name, String genre, int quantity, BigDecimal price){}
+public record DvdStoreDtoWithOption(Long id, String name, String genre, int quantity, float price){}
